requirements.txt
torch>=2.1
torchvision>=0.16
einops>=0.7
timm>=1.0
lpips>=0.1
Pillow>=10.0
